[sim] fix mismatched types in MIL_UrbanObject

+++ src/libraries/simulation_kernel/Urban/MIL_UrbanObject.h
@@ -181,7 +181,7 @@ private:
 public:
     //! @name Member data
     //@{
-    static int maxScaledLocationsNumber_;
+    static size_t maxScaledLocationsNumber_;
     //@}
 };
